requirements-cli.txt
cffi==1.16.0 ; (platform_system == "Darwin" or platform_system == "Windows" or platform_system == "Linux") and platform_python_implementation == "CPython" and python_version >= "3.11" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.12" and platform_system == "Windows"
cryptography==42.0.5 ; (platform_system == "Darwin" or platform_system == "Windows" or platform_system == "Linux") and platform_python_implementation == "CPython" and python_version >= "3.11" and python_version < "3.12"
cssselect==1.2.0 ; python_version >= "3.11" and python_version < "3.12"
ebooklib==0.18 ; python_version >= "3.11" and python_version < "3.12"
h11==0.14.0 ; python_version >= "3.11" and python_version < "3.12"
h2==4.1.0 ; python_version >= "3.11" and python_version < "3.12"
hpack==4.0.0 ; python_version >= "3.11" and python_version < "3.12"
hyperframe==6.0.1 ; python_version >= "3.11" and python_version < "3.12"
idna==3.6 ; python_version >= "3.11" and python_version < "3.12"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "3.12"
kiss-headers==2.4.3 ; python_version >= "3.11" and python_version < "3.12"
linkify-it-py==2.0.3 ; python_version >= "3.11" and python_version < "3.12"
lxml==5.1.0 ; python_version >= "3.11" and python_version < "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "3.12"
markdown-it-py[linkify,plugins]==3.0.0 ; python_version >= "3.11" and python_version < "3.12"
mdit-py-plugins==0.4.0 ; python_version >= "3.11" and python_version < "3.12"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "3.12"
niquests==3.5.3 ; python_version >= "3.11" and python_version < "3.12"
packaging==24.0 ; python_version >= "3.11" and python_version < "3.12"
parsel==1.8.1 ; python_version >= "3.11" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.11" and (platform_system == "Darwin" or platform_system == "Windows" or platform_system == "Linux") and platform_python_implementation == "CPython" and python_version < "3.12"
pygments==2.17.2 ; python_version >= "3.11" and python_version < "3.12"
qh3==0.15.0 ; (platform_system == "Darwin" or platform_system == "Windows" or platform_system == "Linux") and platform_python_implementation == "CPython" and python_version >= "3.11" and python_version < "3.12"
rich==13.7.1 ; python_version >= "3.11" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.11" and python_version < "3.12"
textual==0.52.1 ; python_version >= "3.11" and python_version < "3.12"
tomli==2.0.1 ; python_version >= "3.11" and python_version < "3.12"
tqdm==4.66.2 ; python_version >= "3.11" and python_version < "3.12"
trogon==0.5.0 ; python_version >= "3.11" and python_version < "3.12"
typing-extensions==4.10.0 ; python_version >= "3.11" and python_version < "3.12"
uc-micro-py==1.0.3 ; python_version >= "3.11" and python_version < "3.12"
urllib3-future==2.6.903 ; python_version >= "3.11" and python_version < "3.12"
w3lib==2.1.2 ; python_version >= "3.11" and python_version < "3.12"
wassima==1.1.0 ; python_version >= "3.11" and python_version < "3.12"
